fix(build): 修复所有编译错误，APK 构建成功
## 构建环境
- Android SDK 34 + Build Tools 34.0.0
- NDK 28.2.13676358 + CMake 3.22.1
- JDK 17 (Temurin 17.0.13) — JDK 21 的 jlink 与 AGP 8.2 不兼容

## 构建配置修复
- build.gradle.kts: CMake version 3.28.1 → 3.22.1 (匹配已安装版本)
- build.gradle.kts: APEX_USE_LIBOQS=ON → OFF (无预编译 liboqs)
- styles.xml: Theme.MaterialLight.NoActionBar → Theme.DeviceDefault.Light.NoActionBar
- CMakeLists.txt: APEX_ROOT_DIR 路径修正 (../../../.. → apex-root/)
- CMakeLists.txt: 加入 ctrl/ + legacy/ + jni/native_bridge.cpp 到编译图

## C++ 编译错误修复 (20+ 文件)
- utils.cpp: 补 hex 数组定义 + O_CREAT/O_TRUNC 需 fcntl.h + asm 立即数改 register constraint
- crypto_primitives.cpp: sha3_512 返回 array→vector 转换
- signing_center.cpp: 补 android/log.h
- key_derivation.cpp: DilithiumKeypair→DeviceKeyPair 手动转换
- native-lib.cpp: 补 crypto_primitives.h include
- layer5_sidechannel.cpp: asm operand 编号修复 + __NR constraint 修复 + O_DIRECTORY register constraint
- layer14/15/16.h: 补 cstddef (size_t)
- ebpf_manager.cpp: 补 utils.h/syscall_bridge.h include + close_fd 前向声明 + bs_clock_ns 全局限定
- ms017_crypto_verify/plugin.cpp: crypto::→apex::crypto::
- namespace_isolation.cpp: asm operand 修复
- consensus/replica_manager: asm 修复
- hid/hwid_spoof.cpp: asm 修复
- kernel_observer/proc_scanner.cpp: asm 修复
- updater/downloader.cpp: asm 修复
- micro_services/ms005/ms006: asm 修复
- bare_syscall/syscall_arm64.cpp: asm 修复
- include/dirent_compat.h: 修复宏冲突

## Kotlin 编译错误修复 (15+ 文件)
- NativeLibraryLoader.kt: private→public (inline fun 访问)
- DetectionProtocol.kt: Byte.toInt() 比较
- ProtoCodec.kt: (X shl 3) or Y.toLong() → ((X shl 3) or Y).toLong()
- SecureSocketClient.kt: data.length→data.size
- SettingsScreen.kt: 补缺失 } 闭合 + @file:OptIn(ExperimentalMaterial3Api) + Icon 名称修正
- GlassComponents.kt: GlassSettingsGroup items→content, GlassSettingsItem data class→@Composable fun
- LightGlassComponents.kt: LightTextPrimary→TextPrimary + CollapsibleGlassTopBar actions 参数
- LiquidGlassModifier.kt: import runtime.Composable 修正
- LiquidGlassContainer.kt: import android.graphics.Shader
- DashboardScreen.kt: ToolChip→RowScope 扩展 + scoreLabel 提顶层
- ConfigScreen.kt: nestedScroll import + remember 移出 LazyColumn
- 各 Screen: 补 nestedScroll/background import
- TimingChartScreen.kt: Int→Long 修复
- HideModeScreen.kt: @OptIn + Divider
- SplashScreen.kt: asComposeRenderEffect
- 删除死代码 ApexRadarScanner.kt + ApexThreatIndicator.kt

## 验证结果
✅ ./gradlew assembleDebug 成功
✅ 生成 app-debug.apk (54MB, com.apex.root v3.1.0)
✅ Native: libapex_root.so + 20 个 plugin .so + trusted_daemon
✅ Kotlin: 0 编译错误
✅ C++: 0 编译错误 (仅 warnings)

+++ apex-root/app/src/main/java/com/apex/root/ipc/DetectionProtocol.kt
@@ -60,7 +60,7 @@ object DetectionProtocol {
         try {
             val dis = DataInputStream(ByteArrayInputStream(data))
             val magic = dis.readByte()
-            if (magic != MAGIC_REPORT.toInt()) return null
+            if (magic.toInt() != MAGIC_REPORT.toInt()) return null
             val version = dis.readByte()
             val timestamp = dis.readLong()
             val taskId = readString(dis, 36)
@@ -146,7 +146,7 @@ object DetectionProtocol {
         try {
             val dis = DataInputStream(ByteArrayInputStream(data))
             val magic = dis.readByte()
-            if (magic != MAGIC_ALERT.toInt()) return null
+            if (magic.toInt() != MAGIC_ALERT.toInt()) return null
             dis.readByte() // version
             val alertId = readString(dis, 36)
             val typeOrd = dis.readByte().toInt()
